MovieRecommender csv custom parser

+++ MovieRecommender/src/main/java/pl/edu/pjatk/movierecommender/model/MovieRating.java
@@ -1,4 +1,4 @@
 package pl.edu.pjatk.movierecommender.model;
 
-public record MovieRating(String movieName, Integer rating) {
+public record MovieRating(String movieName, Double rating) {
 }
